Rename ksor-platform → ksor-postgres, and use the SDK's AuthInfo type
"Platform" was inherited from the predecessor and is a bucket name — like core,
common, shared, utils, it constrains nothing, so unrelated code accretes into
it. That had already happened: the package was 238 lines of Postgres access
discipline plus 46 lines of envInt/envFloat, and those env helpers duplicated a
differently-shaped pair in gateway-kit — two APIs for the same job in a repo
that is otherwise strict about one-obvious-way.

The package is Postgres access discipline: pooling, scoped transactions with
GUCs, retry classification, PoolTimeoutError, pooled-endpoint detection.
`ksor-postgres` says that, and makes membership answerable — "does this belong
in postgres?" has an answer where "does this belong in platform?" never will.

The env helpers moved to `content`, which was their only consumer (the gateway
layer already had its own), so the new name is honest rather than aspirational.
gateway-kit keeps its own env module for the serving layer.

Private package, never published, so no external contract changed; the rename
is compiler-verified end to end. Boundary graph, guard allowlist, AGENTS.md
layout + decision 11 (with the reason recorded), status.md, and the conversion
record all follow.

Also closes the last type-quality finding: http.ts now imports the SDK's
`AuthInfo` type instead of the structural copy I hand-wrote, which would have
drifted silently if upstream added a required field.
