+++ packages/postgres/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@panaversity/ksor-platform",
+  "name": "@panaversity/ksor-postgres",
   "version": "0.0.0",
   "private": true,
-  "description": "Shared contracts and database plumbing for the KSoR content kernel — converted from the production sor-platform trim.",
+  "description": "Postgres access discipline for the ksor kernel: pooling, scoped transactions, retry classification",
   "license": "Apache-2.0",
   "author": "Panaversity",
   "type": "module",
