+++ packages/postgres/src/index.ts
@@ -1,4 +1,3 @@
-export { envFloat, envInt } from "./env.js";
 export {
   createPool,
   DB_BACKOFF_S,
